requirements.txt
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==3.4.1
transformers==4.49.0
torch>=1.13,<3.0

python-dotenv
pandas
scikit-learn
scipy
tqdm
openai>=1.2.4,<2.0.0
nltk
beautifulsoup4

flask
flask-cors